requirements.txt
# FastAPI and ASGI Server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Environment Variables
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.3
pydantic[email]==2.5.3

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client (for testing)
httpx==0.26.0

# Development Tools
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
